Viat Block organize Methods
