+++ .gitignore
@@ -57,3 +57,10 @@ Thumbs.db
 # PRIVATE #
 ######################
 *.private
+
+
+
+.nx/cache
+.nx/workspace-data
+.cursor/rules/nx-rules.mdc
+.github/instructions/nx.instructions.md
