+++ eslint.config.js
@@ -11,7 +11,7 @@ const customGlobals = {
 	Client: 'off',
 	client: 'off',
 	clients: 'off',
-	status: 'off'
+	status: 'off',
 };
 function addGlobals(keysObject) {
 	const keys = Object.keys(keysObject);
@@ -24,7 +24,7 @@ const globalsArray = [
 	globals.commonjs,
 	globals.node,
 	globals.serviceworker,
-	globals.worker
+	globals.worker,
 ];
 globalsArray.forEach(addGlobals);
 Object.assign(globalsObject, customGlobals);
@@ -33,12 +33,12 @@ export default [
 		ignores: [
 			'node_modules/*',
 			'.eslintignore',
-			'**/*.mjs'
+			'**/*.mjs',
 		],
 		files: [
 			'**/*.js',
 			'**/*.umm',
-			'**/*.uml'
+			'**/*.uml',
 		],
 		languageOptions: {
 			parser: babelParser,
@@ -51,15 +51,15 @@ export default [
 		},
 		plugins: {
 			jsdoc,
-			'@stylistic': stylisticJs
+			'@stylistic': stylisticJs,
 		},
 		rules: {
 			'@stylistic/array-bracket-newline': [
 				'error',
 				{
 					multiline: true,
-					minItems: 3
-				}
+					minItems: 3,
+				},
 			],
 			'@stylistic/array-bracket-spacing': ['error', 'never'],
 			'@stylistic/array-element-newline': [
@@ -67,9 +67,9 @@ export default [
 				{
 					ArrayExpression: 'consistent',
 					ArrayPattern: {
-						minItems: 2
-					}
-				}
+						minItems: 2,
+					},
+				},
 			],
 			'@stylistic/arrow-parens': ['error', 'always'],
 			'@stylistic/block-spacing': ['error', 'always'],
@@ -77,19 +77,19 @@ export default [
 			'@stylistic/comma-dangle': [
 				'error',
 				{
-					arrays: 'ignore',
-					exports: 'ignore',
-					functions: 'ignore',
-					imports: 'ignore',
-					objects: 'ignore'
-				}
+					objects: 'always-multiline',
+					arrays: 'always-multiline',
+					imports: 'always-multiline',
+					exports: 'always-multiline',
+					functions: 'never',
+				},
 			],
 			'@stylistic/comma-spacing': [
 				'error',
 				{
 					after: true,
-					before: false
-				}
+					before: false,
+				},
 			],
 			'@stylistic/comma-style': ['error', 'last'],
 			'@stylistic/computed-property-spacing': 'error',
@@ -102,8 +102,8 @@ export default [
 				'error',
 				{
 					after: true,
-					before: false
-				}
+					before: false,
+				},
 			],
 			'@stylistic/implicit-arrow-linebreak': ['error', 'beside'],
 			'@stylistic/indent': [
@@ -112,8 +112,8 @@ export default [
 				{
 					ImportDeclaration: 1,
 					ObjectExpression: 1,
-					SwitchCase: 1
-				}
+					SwitchCase: 1,
+				},
 			],
 			'@stylistic/jsx-quotes': ['error', 'prefer-double'],
 			'@stylistic/key-spacing': [
@@ -121,15 +121,15 @@ export default [
 				{
 					afterColon: true,
 					beforeColon: false,
-					mode: 'strict'
-				}
+					mode: 'strict',
+				},
 			],
 			'@stylistic/keyword-spacing': [
 				'error',
 				{
 					after: true,
-					before: true
-				}
+					before: true,
+				},
 			],
 			'@stylistic/linebreak-style': ['error', 'unix'],
 			'@stylistic/lines-around-comment': [
@@ -138,8 +138,8 @@ export default [
 					afterBlockComment: false,
 					afterLineComment: false,
 					beforeBlockComment: false,
-					beforeLineComment: false
-				}
+					beforeLineComment: false,
+				},
 			],
 			'@stylistic/lines-between-class-members': ['error', 'never'],
 			'@stylistic/max-len': [
@@ -152,22 +152,22 @@ export default [
 					ignoreTemplateLiterals: true,
 					ignoreTrailingComments: true,
 					ignoreUrls: true,
-					tabWidth: 2
-				}
+					tabWidth: 2,
+				},
 			],
 			'@stylistic/max-statements-per-line': [
 				'error',
 				{
-					max: 1
-				}
+					max: 1,
+				},
 			],
 			'@stylistic/multiline-ternary': ['error', 'never'],
 			'@stylistic/new-parens': 'error',
 			'@stylistic/newline-per-chained-call': [
 				'error',
 				{
-					ignoreChainWithDepth: 3
-				}
+					ignoreChainWithDepth: 3,
+				},
 			],
 			'@stylistic/no-confusing-arrow': 'error',
 			'@stylistic/no-extra-parens': 'off',
@@ -180,8 +180,8 @@ export default [
 				'error',
 				{
 					max: 0,
-					maxEOF: 1
-				}
+					maxEOF: 1,
+				},
 			],
 			'@stylistic/no-tabs': 'off',
 			'@stylistic/no-trailing-spaces': 'error',
@@ -191,29 +191,28 @@ export default [
 				{
 					ExportDeclaration: {
 						minProperties: 4,
-						multiline: true
+						multiline: true,
 					},
 					ImportDeclaration: {
 						minProperties: 4,
-						multiline: true
+						multiline: true,
 					},
 					ObjectExpression: {
 						minProperties: 1,
-						multiline: true
+						multiline: true,
 					},
 					ObjectPattern: {
 						minProperties: 2,
-						multiline: true
-					}
-				}
+						multiline: true,
+					},
+				},
 			],
 			'@stylistic/object-curly-spacing': ['error', 'always'],
 			'@stylistic/object-property-newline': [
 				'error',
 				{
 					allowAllPropertiesOnSameLine: false,
-					allowMultiplePropertiesPerLine: false
-				}
+				},
 			],
 			'@stylistic/one-var-declaration-per-line': ['error', 'always'],
 			'@stylistic/operator-linebreak': ['error', 'after'],
@@ -224,8 +223,8 @@ export default [
 				'error',
 				'single',
 				{
-					allowTemplateLiterals: true
-				}
+					allowTemplateLiterals: true,
+				},
 			],
 			'@stylistic/rest-spread-spacing': ['error', 'never'],
 			'@stylistic/semi': ['error', 'always'],
@@ -236,22 +235,22 @@ export default [
 				{
 					anonymous: 'never',
 					asyncArrow: 'always',
-					named: 'never'
-				}
+					named: 'never',
+				},
 			],
 			'@stylistic/space-in-parens': ['error', 'never'],
 			'@stylistic/space-infix-ops': [
 				'error',
 				{
-					int32Hint: false
-				}
+					int32Hint: false,
+				},
 			],
 			'@stylistic/space-unary-ops': [
 				'error',
 				{
 					nonwords: false,
-					words: true
-				}
+					words: true,
+				},
 			],
 			'@stylistic/spaced-comment': [
 				'error',
@@ -260,20 +259,20 @@ export default [
 					block: {
 						balanced: true,
 						exceptions: ['*'],
-						markers: ['!']
+						markers: ['!'],
 					},
 					line: {
 						exceptions: ['-', '+'],
-						markers: ['/']
-					}
-				}
+						markers: ['/'],
+					},
+				},
 			],
 			'@stylistic/switch-colon-spacing': [
 				'error',
 				{
 					after: true,
-					before: false
-				}
+					before: false,
+				},
 			],
 			'@stylistic/template-curly-spacing': ['error', 'never'],
 			'@stylistic/template-tag-spacing': ['error', 'always'],
@@ -287,8 +286,8 @@ export default [
 				'error',
 				{
 					after: true,
-					before: true
-				}
+					before: true,
+				},
 			],
 			'block-scoped-var': 'error',
 			'callback-return': 'off',
@@ -310,8 +309,8 @@ export default [
 			'id-length': [
 				'error',
 				{
-					min: 1
-				}
+					min: 1,
+				},
 			],
 			'id-match': 'off',
 			'init-declarations': 'off',
@@ -364,8 +363,8 @@ export default [
 			'line-comment-position': [
 				'error',
 				{
-					position: 'above'
-				}
+					position: 'above',
+				},
 			],
 			'max-depth': 'off',
 			'max-nested-callbacks': ['error', 3],
@@ -447,8 +446,8 @@ export default [
 			'no-param-reassign': [
 				'error',
 				{
-					props: false
-				}
+					props: false,
+				},
 			],
 			'no-path-concat': 'error',
 			'no-plusplus': 'off',
@@ -459,8 +458,8 @@ export default [
 			'no-redeclare': [
 				'error',
 				{
-					builtinGlobals: true
-				}
+					builtinGlobals: true,
+				},
 			],
 			'no-regex-spaces': 'error',
 			'no-restricted-globals': 'error',
@@ -477,8 +476,8 @@ export default [
 				'error',
 				{
 					builtinGlobals: true,
-					hoist: 'all'
-				}
+					hoist: 'all',
+				},
 			],
 			'no-shadow-restricted-names': 'error',
 			'no-spaced-func': 'error',
@@ -494,8 +493,8 @@ export default [
 			'no-underscore-dangle': [
 				'error',
 				{
-					allow: ['__dirname', '__filename']
-				}
+					allow: ['__dirname', '__filename'],
+				},
 			],
 			'no-unexpected-multiline': 'error',
 			'no-unmodified-loop-condition': 'error',
@@ -505,8 +504,8 @@ export default [
 				'error',
 				{
 					allowShortCircuit: true,
-					allowTernary: true
-				}
+					allowTernary: true,
+				},
 			],
 			'no-unused-labels': 'error',
 			'no-unused-vars': 'off',
@@ -543,9 +542,9 @@ export default [
 						'none',
 						'all',
 						'multiple',
-						'single'
-					]
-				}
+						'single',
+					],
+				},
 			],
 			'sort-keys': 'off',
 			'sort-vars': 'off',
@@ -553,7 +552,7 @@ export default [
 			'use-isnan': 'error',
 			'valid-typeof': 'error',
 			'vars-on-top': 'error',
-			yoda: 'off'
-		}
-	}
+			yoda: 'off',
+		},
+	},
 ];
