+++ viat/blocks/methods/filesystem.js
@@ -1,0 +1,10 @@
+import path from 'path';
+import { write } from '#utilities/file';
+const methods = {
+	async save(directoryPath) {
+		const blockBinary = await this.exportBinary();
+		const fullSavePath = path.join(directoryPath, await this.getPath());
+		return write(fullSavePath, blockBinary, 'binary', true);
+	},
+};
+export default methods;
