+++ viat/blocks/wallet/block.js
@@ -49,10 +49,11 @@ export async function walletBlock(...args) {
 	return block;
 }
 export default walletBlock;
-const amy = await wallet();
-const amyBlock = await walletBlock(amy);
-await amyBlock.finalize();
-console.log('Wallet Block', amyBlock.block);
+// const amy = await wallet();
+// const amyBlock = await walletBlock(amy);
+// await amyBlock.finalize();
+// await amyBlock.sign(amy);
+// console.log('Wallet Block', amyBlock.block);
 // const exampleBlock = await walletBlock({
 // 	sender: viatCipherSuite.createBlockNonce(64),
 // 	transaction: viatCipherSuite.createBlockNonce(64),
