+++ viat/wallet/wallet.js
@@ -24,7 +24,7 @@ export class Wallet extends CryptoID {
 		});
 		await txBlock.finalize();
 		await txBlock.sign(this);
-		await txBlock.createReceipt();
+		await txBlock.setReceipt();
 		// await txBlock.receipt.finalize();
 		// await txBlock.receipt.sign(this);
 		console.log('Transaction Block:', txBlock.block);
